Added polygon to the map when fetching data
still does not work with parsing multiPolygon areas from the api like Italy which has islands

+++ app/src/main/java/com/col/eventradar/models/LocationSearchResult.kt
@@ -1,7 +1,7 @@
 package com.col.eventradar.models
 
 data class LocationSearchResult(
-    val osmId: Double,
+    val osmId: Long,
     val locationName: String,
     val latitude: Double,
     val longitude: Double,
